Fix negative profile time in system() and fork().
The calculation of time used and remaining in itimers was incorrect.
Use getitimer() calls to get time spent; setitimer() returns the old
counter setting, not how much is left on the ticking counter.

Also fixes a bug where the timers could be left on even if they were
off on entry, for example turned off in the exit path code.