First version of igprof-book.
igprof-book is a redesign of igprof-navigator. The main change is that
it can now handle a number of igprof dumps at the same time, which
allows for correlation / diffing studies. The web GUI itself has been
largely improved and it is now much more dynamic than before, allowing
for sorting of columns.

Finally there is a nicer entry page which shows historical data being
kept in the database.