igprof-analyse -ma/--merge-ancestors "REGEX1>REGEX2>...>REGEXN/NEW_NAME"
The -ma option go trough nodes in the call tree and uses PCRE check if
node is child of ancestors that matches to REGEX list.
Node that matches REGEXN and is child of all ancestors that matches previos
regular expressions will renamed as NEW_NAME. Also all child of the node
REGEXN will get the name NEW_NAME

Ancestors got to match in order. REGEX1 is the oldest and REGEX2 is the 2nd
oldest.

For example program has function foo which calls bar and bar calls foobar
and at some other place the program calls only bar which then calls foobar.
-ma "foo>bar>foobar/foobar: child of foo and bar" would rename foobar at
the first case but leave foobar unchanged at the second case

Node doesn't need to be direct child of ancestor
For example -ma "foo>foobar/foobar: has foo ancestor" Would rename foobar
if foo is ancestor of the foobar.