Fix crash on ARM when running performance profiling.
Looks like the dynamic loader invokes `close` on ARM, which leads to a
segfault in the doclose method. For the moment I just exclude the hook,
however given it is actually used to protect an output corruption in case a
program closes stderr, we should probably find a better solution.