monkeypatch dl_iterate_phdr to avoid deadlocks
dl_iterate_phdr can be called by the program we're profiling (e.g.,
by libc). When a sample fires in while dl_iterate_phdr has locked
its mutex, we run our own dl_iterate_phdr that then attempts to
lock the same mutex, resulting in a deadlock. In order to prevent
that, we override the definition of dl_iterate_phdr and protect
the original definition so it cannot be called within the same
thread if it's already being called.

Attempts at doing this with LIBHOOK and DUAL_HOOK were unsuccessful.